lisp: factor out metadata for user commands into a separate table.
Use the lua function itself as the key into the metadata ready for
the transition to extending with pure Lua instead of interpreted
lisp.
* src/lisp.lua (usercmd): Rename from this...
(zi): ...to this. Adjust all callers.
(introspect): A table with weak keys for Defun metadata.
(Defun): Make the value for a Defun named in the `zi' sandbox a
callable function.  Put the `doc' and `interactive' metadata in
the new introspect table.  Adjust all callers.