keycode.lua: simplify handling alternative names for keys, and fix a bug
There was a Heisenbug where different stringifications of "\\M- " as
either {key=32,META=true} or {META=true,key=32} would give different
memoized keycodes, and hence one form would not find a keybinding (for
just-one-space in the current setup).

To fix this, ensure that the keystring is canonicalized before calling
the memoized function.

In the process, simplify the handling of keys with more than one name.