zlisp: rewrite the string tokenizer.
* zmacs/eval.lua (read_char): Rename...
(nextch): ...to this, and simplify.  The `i' parameter is now the
index of the next unparsed character in `s', with an imaginary
EOF character at index `#s + 1', so the returned index doesn't need
a special condition anymore.
(read_token): Rename...
(lex): ...to this. Rewritten to accomodate the `i' invariant above,
but also returning "string" and "incomplete string" token kinds
for better diagnostics.  Several other improvements, including
discarding multiple consecutive comments and whitespace on a single
call.
(lisp_read): Rename...
(parse): ...to this. Adjust to work with improved lex above.