completion: fix TAB completion to shortest common prefix (savannah #35112)
* src/completion.lua (completion_try): When building the list of
potential completions, only insert matches that are at least as
long as the search string.  Previously, a search that has both a
longer and a shorter match on the common substring of all three,
would be truncated to the shortest of all!

Fix-Bug: https://savannah.gnu.org/bugs/index.php?35112